TRIPLEX-852 Page: AI-Ready (Page-ai.md, modern stories, JSDoc + тесты)
